added stream function calculation and coloring

+++ src/presets.rs
@@ -178,14 +178,14 @@ pub fn cylinder_cross_flow() -> SpaceTimeDomainPreset {
     let radius = 5.0;
     let center = [20, 20];
     
-    for xi in 14..26 {
-        for yi in 14..26 {
+    for xi in 14..26 as i32 {
+        for yi in 14..26 as i32 {
             let x_dist = xi - center[0];
             let y_dist = yi - center[1];
             let distance = ((x_dist * x_dist + y_dist * y_dist) as f32 ).sqrt();
     
             if distance <= radius {
-                space_domain[xi][yi] = Cell {
+                space_domain[xi as usize][yi as usize] = Cell {
                     cell_type: CellType::BoundaryConditionCell(BoundaryConditionCell::NoSlipCell),
                     ..Default::default()
                 };
